# packages/shared/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ASEP Engine",
    "description": "AI engine for the ASEP platform: chat, agents, repository intelligence.",
    "version": "0.1.0"
  },
  "paths": {
...
    "/v1/integrations/{kind}": {
      "put": {
        "summary": "Set Integration",
        "operationId": "set_integration_v1_integrations__kind__put",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ConnectionIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ConnectionOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "summary": "Delete Integration",
        "operationId": "delete_integration_v1_integrations__kind__delete",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/v1/integrations/{kind}/test": {
      "post": {
        "summary": "Test Integration",
        "description": "Send a test message now, so the settings page can prove it works.\n\nSlack only \u2014 a tracker's \"test\" would create a junk issue, so pushing a real\nwork item is its own proof instead (EXTERNAL_INTEGRATIONS.md).",
        "operationId": "test_integration_v1_integrations__kind__test_post",
        "parameters": [
          {
            "name": "kind",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Kind"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TestResult"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ConnectionIn": {
        "properties": {
          "config": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Config"
          }
        },
        "type": "object",
        "required": [
          "config"
        ],
        "title": "ConnectionIn"
      },
      "ConnectionOut": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "label",
          "enabled",
          "updated_at"
        ],
        "title": "ConnectionOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "TestResult": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "dry_run": {
            "type": "boolean",
            "title": "Dry Run"
          },
          "detail": {
            "type": "string",
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "ok",
          "dry_run",
          "detail"
        ],
        "title": "TestResult"
      },
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}